Add cognate dataset to materials for lab 3
I added the cognate dataset, but I do not understand how to get it in
the right format such that it can be processed by phangorn/ape. I think
it would be a good idea to preprocess it and then deliver it to the
students already in the correct format.